Fix review issues: * move data errors types to own folder * remove redundant comments from data error types * Fix typos
