+++ shared_model/interfaces/query_responses/error_responses/no_roles_error_response.hpp
@@ -15,35 +15,27 @@
  * limitations under the License.
  */
 
-#ifndef IROHA_SHARED_MODEL_STATELESS_FAILED_ERROR_RESPONSE_HPP
-#define IROHA_SHARED_MODEL_STATELESS_FAILED_ERROR_RESPONSE_HPP
+#ifndef IROHA_SHARED_MODEL_NO_ROLES_ERROR_RESPONSE_HPP
+#define IROHA_SHARED_MODEL_NO_ROLES_ERROR_RESPONSE_HPP
 
 #include "interfaces/common_objects/types.hpp"
-#include "interfaces/query_responses/abstract_error_response.hpp"
+#include "interfaces/query_responses/error_responses/abstract_error_response.hpp"
 #include "utils/string_builder.hpp"
 
 namespace shared_model {
   namespace interface {
     /**
-     * Error response of broken query's stateless validation
+     * Error response of broken query, no specified roles
      */
-    class StatelessFailedErrorResponse
-        : public AbstractErrorResponse<StatelessFailedErrorResponse> {
+    class NoRolesErrorResponse
+        : public AbstractErrorResponse<NoRolesErrorResponse> {
      private:
-      /**
-       * @return stateless failed error as error reason
-       */
-      std::string reason() const override {
-        return "StatelessFailedErrorResponse";
-      }
+      std::string reason() const override { return "NoRolesErrorResponse"; }
 
-      /**
-       * @return Old model error reason.
-       */
       iroha::model::ErrorResponse::Reason oldModelReason() const override {
-        return iroha::model::ErrorResponse::Reason::STATELESS_INVALID;
+        return iroha::model::ErrorResponse::Reason::NO_ROLES;
       }
     };
   }  // namespace interface
 }  // namespace shared_model
-#endif  // IROHA_SHARED_MODEL_STATELESS_FAILED_ERROR_RESPONSE_HPP
+#endif  // IROHA_SHARED_MODEL_NO_ROLES_ERROR_RESPONSE_HPP
